refactor: arruma diabetes no frontend para usar a estrutura de modules
