+++ src/app/lista-nominal/page.tsx
@@ -1,3 +1,3 @@
-import { AcfPage } from "@/features/acf/frontend/diabetes/AcfPage";
+import { AcfPage } from "@/features/acf/frontend/diabetes";
 
 export default AcfPage;
