+++ src/features/acf/frontend/common/WithFilters/model.ts
@@ -1,5 +1,5 @@
 //TODO: Pensar melhor nesse import, tá muito nested.
-import type * as diabetes from "@/features/acf/frontend/diabetes/AcfPage/modules/PanelSelector/modules/List/modules/CoeqDataTable";
+import type * as diabetes from "@/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/CoeqDataTable";
 
 /// Contém todos os possíveis conjuntos de filtros.
 /// Eventualmente, este Union type deve ter um item para cada combinação indicador x perfil
