+++ src/features/acf/frontend/diabetes/index.tsx
@@ -1,0 +1,1 @@
+export { AcfPage } from "./modules/AcfPage";
