+++ src/features/acf/frontend/diabetes/modules/AcfPage/index.tsx
@@ -4,7 +4,7 @@ import { SessionGuard } from "@/features/common/frontend/SessionGuard";
 import type { ProfileIdValue } from "@/types/profile";
 import { PROFILE_ID } from "@/types/profile";
 import { getServerSession } from "next-auth";
-import type { AcfDashboardType } from "../../../shared/diabetes/model";
+import type { AcfDashboardType } from "../../../../shared/diabetes/model";
 import { ErrorPage } from "./modules/ErrorPage";
 import { PanelSelector } from "./modules/PanelSelector";
